requirements.txt
librosa==0.6.2
matplotlib==3.0.2
numpy==1.14.5
scipy==1.1.0
tensorflow-gpu==1.10.0
Unidecode==0.4.20
